Remove unneeded scripts for d8, r8 and retrace
The wrappers used to hardcode a memory setting of 2g, but with later
jdk releases the max memory setting has increased and there should not
be a need to hard code the setting.

The retracing of R8 stack traces can be done with the new 'retrace'.

The 'd8-compat-dx' script is removed since the target was removed in
commit 1f6a318a0373fa6a304c3c581c9b3a5583054224.

Bug: b/227746536
Test: n/a
Change-Id: I25e8d6e2565c1c862dc5e731aab2435e708f8a11
Signed-off-by: Jason Edson <jaysonedson@gmail.com>